implement correct fix for segfault on 64bit platform
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@209957 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_compile.c
@@ -4377,7 +4377,7 @@ void zend_do_label(znode *label TSRMLS_DC)
 void zend_resolve_goto_label(zend_op_array *op_array, zend_op *opline, int pass2 TSRMLS_DC)
 {
 	zend_label *dest;
-	int current, distance;
+	long current, distance;
 
 	if (CG(labels) == NULL ||
 	    zend_u_hash_find(CG(labels), Z_TYPE(opline->op2.u.constant), Z_UNIVAL(opline->op2.u.constant), Z_UNILEN(opline->op2.u.constant)+1, (void**)&dest) == FAILURE) {
